// workspaces/liferay-customer-workspace/client-extensions/liferay-customer-custom-element/src/common/interfaces/accountBrief.ts
/**
 * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

import IRoleBrief from './rolelBrief';

export default interface IAccountBrief {
	externalReferenceCode: string;
	id: number;
	name: string;
	roleBriefs: IRoleBrief[] | undefined;
}